+++ packages/cashscript/src/index.ts
@@ -3,6 +3,7 @@ import SignatureTemplate from './SignatureTemplate.js';
 export { SignatureTemplate };
 export { Contract, ContractFunction } from './Contract.js';
 export { Transaction } from './Transaction.js';
+export { TransactionBuilder } from './TransactionBuilder.js';
 export { Argument, encodeArgument } from './Argument.js';
 export { Artifact, AbiFunction, AbiInput } from '@cashscript/utils';
 export * as utils from '@cashscript/utils';
@@ -12,7 +13,7 @@ export {
   SignatureAlgorithm,
   HashType,
   Network,
-  isSignableUtxo,
+  isUtxoP2PKH,
 } from './interfaces.js';
 export * from './Errors.js';
 export {